migrator/requirements.txt
nats-py
python-dotenv
dataclasses-json
dacite
aiohttp
Jinja2
faker
psycopg2-binary
requests
